+++ crates/entrypoint/src/cli/command/azure_devops/license_entitlement/user/mod.rs
@@ -2,11 +2,13 @@ mod azure_devops_license_entitlement_user_matcher;
 pub mod list;
 pub mod revoke;
 pub mod show;
+pub mod summary;
 pub mod update;
 
 pub use azure_devops_license_entitlement_user_matcher::*;
 pub use list::AzureDevOpsLicenseEntitlementUserListArgs;
 pub use revoke::AzureDevOpsLicenseEntitlementUserRevokeArgs;
 pub use show::AzureDevOpsLicenseEntitlementUserShowArgs;
+pub use summary::AzureDevOpsLicenseEntitlementUserSummaryArgs;
 pub use update::AzureDevOpsLicenseEntitlementUserUpdateArgs;
 pub use update::AzureDevOpsLicenseEntitlementUserUpdateTuiArgs;