+++ crates/entrypoint/src/cli/command/azure_devops/license_entitlement/user/summary/mod.rs
@@ -1,0 +1,3 @@
+mod azure_devops_license_entitlement_user_summary_cli;
+
+pub use azure_devops_license_entitlement_user_summary_cli::*;